+++ library/src/main/java/ru/otus/domain/service/repository/BookRepository.java
@@ -1,10 +1,10 @@
-package ru.otus.domain.service.dao;
+package ru.otus.domain.service.repository;
 
 import ru.otus.domain.model.Book;
 
 import java.util.List;
 
-public interface BookDao {
+public interface BookRepository {
 
 	List<Book> getAll();
